build: fix pointer-to-integer truncation on 64-bit platforms
Adds intptr_t/uintptr_t intermediate casts in void* → integer conversions
to silence -Wint-conversion and prevent truncation on 64-bit builds.
Renames local variable 'pause' to 'shouldPause' to avoid conflict with
POSIX pause() on Linux.

+++ Generals/Code/GameEngine/Source/Common/Thing/ThingTemplate.cpp
@@ -430,7 +430,8 @@ void ThingTemplate::parseModuleName(INI* ini, void *instance, void* store, const
 {
 	ThingTemplate* self = (ThingTemplate*)instance;
 	ModuleInfo* mi = (ModuleInfo*)store;
-	ModuleType type = (ModuleType)(UnsignedInt)userData;
+	// GeneralsX @build Fix pointer-to-integer truncation on 64-bit platforms.
+	ModuleType type = (ModuleType)(UnsignedInt)(uintptr_t)userData;
 	const char* token = ini->getNextToken();
 	AsciiString tokenStr = token;
 
@@ -531,7 +532,8 @@ void ThingTemplate::parseModuleName(INI* ini, void *instance, void* store, const
 //-------------------------------------------------------------------------------------------------
 void ThingTemplate::parseIntList(INI* ini, void *instance, void* store, const void* userData)
 {
-	Int numberEntries = (Int)userData;
+	// GeneralsX @build Fix pointer-to-integer truncation on 64-bit platforms.
+	Int numberEntries = (Int)(uintptr_t)userData;
 	Int *intList = (Int*)store;
 
 	for( Int intIndex = 0; intIndex < numberEntries; intIndex ++ )
